+++ src/msrv.rs
@@ -1,5 +1,5 @@
-use crate::rust::RustRelease;
 use crate::rust::Toolchain;
+use crate::rust::ToolchainCandidate;
 
 /// An enum to represent the minimal compatibility
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -14,13 +14,13 @@ pub enum MinimumSupportedRustVersion {
 }
 
 impl MinimumSupportedRustVersion {
-    pub fn toolchain(msrv: &RustRelease) -> Self {
+    pub fn toolchain(msrv: &ToolchainCandidate) -> Self {
         let toolchain = msrv.to_toolchain_spec().to_owned();
 
         Self::Toolchain { toolchain }
     }
 
-    pub fn from_option(msrv: Option<&RustRelease>) -> Self {
+    pub fn from_option(msrv: Option<&ToolchainCandidate>) -> Self {
         msrv.map_or(
             MinimumSupportedRustVersion::NoCompatibleToolchain,
             MinimumSupportedRustVersion::toolchain,
@@ -30,7 +30,7 @@ impl MinimumSupportedRustVersion {
 
 impl MinimumSupportedRustVersion {
     #[cfg(test)]
-    pub fn unwrap_version(&self) -> rust_releases::semver::Version {
+    pub fn unwrap_version(&self) -> semver::Version {
         if let Self::Toolchain { toolchain, .. } = self {
             return toolchain.version().clone();
         }
@@ -42,18 +42,14 @@ impl MinimumSupportedRustVersion {
 #[cfg(test)]
 mod tests {
     use crate::msrv::MinimumSupportedRustVersion;
-    use crate::rust::RustRelease;
-    use cargo_metadata::semver;
+    use crate::rust::{RustRelease, Stable, ToolchainCandidate};
 
     #[test]
     fn accept() {
         let version = semver::Version::new(1, 2, 3);
-        let rust_release = RustRelease::new(
-            rust_releases::Release::new_stable(version.clone()),
-            "x",
-            &[],
-        );
-        let msrv = MinimumSupportedRustVersion::toolchain(&rust_release);
+        let candidate =
+            ToolchainCandidate::new(RustRelease::new(Stable::new(1, 2, 3), None, []), "x", &[]);
+        let msrv = MinimumSupportedRustVersion::toolchain(&candidate);
 
         assert!(matches!(
             msrv,
@@ -63,12 +59,9 @@ mod tests {
     #[test]
     fn accept_from_option() {
         let version = semver::Version::new(1, 2, 3);
-        let rust_release = RustRelease::new(
-            rust_releases::Release::new_stable(version.clone()),
-            "x",
-            &[],
-        );
-        let msrv = MinimumSupportedRustVersion::from_option(Some(&rust_release));
+        let candidate =
+            ToolchainCandidate::new(RustRelease::new(Stable::new(1, 2, 3), None, []), "x", &[]);
+        let msrv = MinimumSupportedRustVersion::from_option(Some(&candidate));
 
         assert!(matches!(
             msrv,